refactor: change Fetch parameters from Integer to Long for better precision

+++ plugins/plugin-mysql/src/main/java/io/github/auspis/fluentsql4j/plugin/builtin/mysql/ast/visitor/ps/strategy/MySqlFetchPsStrategy.java
@@ -16,8 +16,8 @@ public class MySqlFetchPsStrategy implements FetchPsStrategy {
         if (!clause.isActive()) {
             return new PreparedStatementSpec("", List.of());
         }
-        Integer offset = clause.offset();
-        Integer rows = clause.rows();
+        Long offset = clause.offset();
+        Long rows = clause.rows();
         String sql = String.format("LIMIT %s OFFSET %s", rows, offset);
         return new PreparedStatementSpec(sql, List.of());
     }
